feat(vision): modularize pipelines
