+++ Server/core/vision/api.py
@@ -1,4 +1,4 @@
-"""Robot vision API backed by a stateful VisionEngine."""
+"""Robot vision API backed by modular pipelines."""
 
 from __future__ import annotations
 
@@ -6,40 +6,43 @@ from typing import Any, Dict, Optional, TYPE_CHECKING
 
 import numpy as np
 
-from .engine import VisionEngine, EngineResult
+from .pipeline import BasePipeline, ContourPipeline, FacePipeline, Result
 
 if TYPE_CHECKING:  # pragma: no cover - for type checkers only
     from .viz_logger import VisionLogger
 
-_ENGINE: Optional[VisionEngine] = None
+_PIPELINES: Dict[str, BasePipeline] = {
+    "object": ContourPipeline(),
+    "face": FacePipeline(),
+}
+_CURRENT: str = "object"
 
 
-def _engine() -> VisionEngine:
-    """Lazily create and return the single VisionEngine instance."""
-    global _ENGINE
-    if _ENGINE is None:
-        _ENGINE = VisionEngine()
-    return _ENGINE
+def _pipeline() -> BasePipeline:
+    return _PIPELINES[_CURRENT]
 
 
 def reset_state() -> None:
     """Reset internal stability state."""
-    _engine().reset_state()
+    _pipeline().reset_state()
 
 
 def load_profile(which: str, path: Optional[str] = None) -> None:
     """Reload a profile ('big' or 'small') and reset state."""
-    _engine().load_profile(which, path)
+    _pipeline().load_profile(which, path)
 
 
 def update_dynamic(which: str, params: Dict[str, Any]) -> None:
     """Update dynamic adjuster parameters at runtime."""
-    _engine().update_dynamic(which, params)
+    _pipeline().update_dynamic(which, params)
 
 
 def select_detector(mode: str) -> None:
-    """Select detector mode."""
-    _engine().select_detector(mode)
+    """Select vision pipeline mode."""
+    global _CURRENT
+    if mode not in _PIPELINES:
+        raise ValueError("mode must be 'object' or 'face'")
+    _CURRENT = mode
 
 
 def process_frame(
@@ -47,28 +50,22 @@ def process_frame(
     return_overlay: bool = True,
     config: Optional[Dict[str, Any]] = None,
 ):
-    """Process ``frame`` and return a detection dict.
-
-    Signature remains compatible with previous versions.
-    """
-    res: EngineResult = _engine().process(
-        frame, return_overlay=return_overlay, config=config
-    )
+    """Process ``frame`` and return a detection dict."""
+    cfg = dict(config or {})
+    cfg["return_overlay"] = return_overlay
+    res: Result = _pipeline().process(frame, cfg)
     return res.data
 
 
 # VisionLogger fetches the latest detection result through this helper.
-def get_last_result() -> Optional[EngineResult]:
-    """Return the most recent EngineResult produced by ``process_frame()``.
-
-    Returns ``None`` if no frame has been processed.
-    """
-    return _engine().get_last_result()
+def get_last_result() -> Optional[Result]:
+    """Return the most recent Result produced by :func:`process_frame`."""
+    return _pipeline().get_last_result()
 
 
 def get_detectors():
     """Return underlying detectors for inspection in tests."""
-    return _engine().get_detectors()
+    return _pipeline().get_detectors()
 
 
 def create_logger_from_env() -> Optional["VisionLogger"]:
